+++ rust/src/years/y2023/mod.rs
@@ -2,7 +2,7 @@ mod day01;
 use crate::utils;
 
 
-pub fn run_day(day: &mut u8, test: bool) -> (i32, i32) {
+pub fn run_day(day: &mut u8, test: bool) -> (i64, i64) {
     let contents = utils::get_day_input_content(day, 2023, test);
 
     println!("Running day {day} of year 2023");
